requirements.txt
pandas>=2.0
SQLAlchemy>=2.0
psycopg2-binary>=2.9
python-dotenv>=1.0
pytest>=8.0
requests>=2.31
